Create README.md
The Sparks Foundation - Machine Learning Tasks
This repository consist of the tasks given during my internship at The Sparks Foundation.

About The Sparks Foundation
The Sparks Foundation mission is to inspire students, help them innovate, and let them integrate to build the next generation humankind. To inspire, motivate, and encourage students to learn, create, and help build a better society. To teach new ways of thinking, to innovate and solve the problems on their own. We help the students to integrate and help each other, learn from each other, and do well together.

TASKS
Task 1 - Exploring Supervised Machine Learning
In this regression task we will predict the percentage of marks that a student is expected to score based upon the number of hours they studied.This is a simple linear regression task as it involves just two variables.Data can be found at http://bit.ly/w-data. What will be predicted score if a student study for 9.25 hrs in a day?

Task 2 - Exploring Unsupervised Machine Learning
From the given 'Iris' dataset, predict the optimum number of clusters and represent it visually.

Task 3 - Exploring Decision Tree Algorithm
For the given 'Iris' dataset, create the Decision Tree classifier and visualize it graphically. The purpose is if we feed any new data to this classifier, it would be able to predict the right class accordingly.